+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.5.6-SNAPSHOT]
+
+- Padding shortcuts (paddingRight, paddingLeft, paddingTop, paddingBottom) to Region
+- TableView support for Nested Columns (`nestedColumn(title) { // add child columns here }`)
+- TableView support for expanded row node (`rowExpander { // create node to show on expand here }`)
+
 ## [1.5.5]
 
 - Stylesheets can be loaded via ServiceLoader (`META-INF/services/tornadofx.Stylesheet` with reference to the stylesheet class)